cleanup: remove enzyme and all the deps, remove all example tests, move compoent into its own folder with test file
