+++ client/src/index.tsx
@@ -1,5 +1,5 @@
 import ReactDOM from 'react-dom';
 import React from 'react';
-import App from './components/App';
+import App from './components/App/App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
