+++ package.json
@@ -34,7 +34,6 @@
     "@babel/preset-env": "^7.13.10",
     "@babel/preset-react": "^7.12.13",
     "@babel/preset-typescript": "^7.13.0",
-    "@types/enzyme": "^3.10.8",
     "@types/jest": "^26.0.21",
     "@types/node": "^14.14.34",
     "@types/react-dom": "^17.0.2",
@@ -42,8 +41,6 @@
     "@typescript-eslint/parser": "^4.19.0",
     "babel-jest": "^26.6.3",
     "babel-loader": "^8.2.2",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.6",
     "eslint": "^7.22.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-plugin-import": "^2.22.1",
@@ -52,7 +49,6 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "jest": "^26.6.3",
     "jest-environment-enzyme": "^7.1.2",
-    "jest-enzyme": "^7.1.2",
     "ts-jest": "^26.5.4",
     "ts-loader": "^8.0.18",
     "ts-node": "^9.1.1",
