--- sum.ts
@@ -1,6 +1,0 @@
-// test jest
-function sum(a: number, b: number): number {
-  return a + b;
-}
-
-export default sum;
