--- test/sum.test.ts
@@ -1,6 +1,0 @@
-// // testing jest
-import sumT from '../sum';
-
-test('adds 1 + 2 to equal 3', () => {
-  expect(sumT(1, 2)).toBe(3);
-});
